requirements.txt
aiogram==X.Y.Z
python-dotenv==X.Y.Z
